Create Session objects
