+++ tests/test_codecs.py
@@ -1,6 +1,6 @@
 # coding: utf-8
+from coreapi import negotiate_decoder, negotiate_encoder
 from coreapi.codecs import CoreJSONCodec, HTMLCodec
-from coreapi.codecs import negotiate_decoder, negotiate_encoder
 from coreapi.codecs.corejson import _document_to_primative, _primative_to_document
 from coreapi.document import Document, Link, Error, required
 from coreapi.exceptions import ParseError, NotAcceptable
